moves consoluser to ee dir
