+++ ee/consoleuser/consoleuser_darwin.go
