+++ ee/consoleuser/consoleuser_test.go
